pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]
